# src/models/dto/PersonDTO.ts
export class PersonDTO {
    id:number;
    name:string;
    email:string;

    constructor(id?:number, name?:string, email?: string) {
        this.id = id || 0;
        this.name = name || '';
        this.email = email || '';
    }
}